+++ api/ExtensionAPIController.php
@@ -1,5 +1,7 @@
 <?php
 
+use Pastell\Service\Droit\DroitService;
+
 class ExtensionAPIController extends BaseAPIController
 {
     public function __construct(
@@ -14,7 +16,7 @@ class ExtensionAPIController extends BaseAPIController
      */
     public function get(): array
     {
-        $this->checkDroit(0, 'system:lecture');
+        $this->checkDroit(0, DroitService::getDroitLecture(DroitService::DROIT_SYSTEM));
         $id_extension = $this->getFromQueryArgs(0);
         if ($id_extension) {
             if (! $this->extensionSQL->getInfo($id_extension)) {
@@ -37,7 +39,7 @@ class ExtensionAPIController extends BaseAPIController
      */
     public function post(): array
     {
-        $this->checkDroit(0, 'system:edition');
+        $this->checkDroit(0, DroitService::getDroitEdition(DroitService::DROIT_SYSTEM));
         $path = $this->getFromRequest('path');
         if (! file_exists($path)) {
             throw new Exception("Le chemin « $path » n'existe pas sur le système de fichier");
@@ -62,7 +64,7 @@ class ExtensionAPIController extends BaseAPIController
      */
     public function patch()
     {
-        $this->checkDroit(0, 'system:edition');
+        $this->checkDroit(0, DroitService::getDroitEdition(DroitService::DROIT_SYSTEM));
         $id_extension = $this->getFromQueryArgs(0);
         if (! $id_extension || ! $this->extensionSQL->getInfo($id_extension)) {
             throw new NotFoundException("Extension #$id_extension non trouvée");
@@ -92,7 +94,7 @@ class ExtensionAPIController extends BaseAPIController
      */
     public function delete()
     {
-        $this->checkDroit(0, "system:edition");
+        $this->checkDroit(0, DroitService::getDroitEdition(DroitService::DROIT_SYSTEM));
         $id_extension = $this->getFromQueryArgs(0);
         if (! $id_extension || ! $this->extensionSQL->getInfo($id_extension)) {
             throw new NotFoundException("Extension #$id_extension non trouvée");
@@ -110,7 +112,7 @@ class ExtensionAPIController extends BaseAPIController
      */
     public function compatV1Edition()
     {
-        $this->checkDroit(0, "system:edition");
+        $this->checkDroit(0, DroitService::getDroitEdition(DroitService::DROIT_SYSTEM));
 
         $id_extension = $this->getFromRequest('id_extension');
         $path = $this->getFromRequest('path');