+++ api/RoleAPIController.php
@@ -1,10 +1,15 @@
 <?php
 
+use Pastell\Service\Droit\DroitService;
+
 class RoleAPIController extends BaseAPIController
 {
+    /**
+     * @throws ForbiddenException
+     */
     public function get()
     {
-        $this->checkOneDroit("role:lecture");
+        $this->checkOneDroit(DroitService::getDroitLecture(DroitService::DROIT_ROLE));
         return $this->getRoleUtilisateur()->getAuthorizedRoleToDelegate($this->getUtilisateurId());
     }
 }