+++ connecteur-type/utilities/GetModuleList.php
@@ -1,18 +1,20 @@
 <?php
 
+use Pastell\Service\Module\ModuleListService;
+
 class GetModuleList extends ConnecteurTypeChoiceActionExecutor
 {
-    private const MODULE_TYPE_FIELD = 'module_type';
-    private const MODULE_TYPE_LABEL_FIELD = 'module_type_label';
-    private const PAGE_TITLE = 'page_title';
+    private const string MODULE_TYPE_FIELD = 'module_type';
+    private const string MODULE_TYPE_LABEL_FIELD = 'module_type_label';
+    private const string PAGE_TITLE = 'page_title';
 
     /**
      * @return bool
      * @throws RecoverableException
      */
-    public function go()
+    public function go(): bool
     {
-        $moduleType = $this->getRecuperateur()->get(self::MODULE_TYPE_FIELD);
+        $moduleType = (string)$this->getRecuperateur()->get(self::MODULE_TYPE_FIELD);
         $moduleList = $this->displayAPI();
         if ($moduleType && empty($moduleList[$moduleType])) {
             throw new RecoverableException("Ce type de dossier n'existe pas");
@@ -31,22 +33,17 @@ class GetModuleList extends ConnecteurTypeChoiceActionExecutor
     /**
      * @throws NotFoundException
      */
-    public function display()
+    public function display(): true
     {
         $this->setViewParameter(
             'moduleType',
             $this->getConnecteurProperties()->get($this->getMappingValue(self::MODULE_TYPE_FIELD))
         );
-        $modules = $this->displayAPI();
 
-        $currentLocale = setlocale(LC_COLLATE, '0');
-        setlocale(LC_COLLATE, 'fr_FR.utf8');
-        uasort($modules, static function (array $a, array $b) {
-            return strcoll($a['nom'], $b['nom']);
-        });
-        setlocale(LC_COLLATE, $currentLocale);
-
-        $this->setViewParameter('moduleList', $modules);
+        $this->setViewParameter(
+            'moduleList',
+            $this->objectInstancier->getInstance(ModuleListService::class)->getModuleListOrderByType($this->id_u)
+        );
         $this->renderPage(
             $this->getMappingValue(self::PAGE_TITLE),
             'connectorType/utilities/GetModuleList'
@@ -54,8 +51,10 @@ class GetModuleList extends ConnecteurTypeChoiceActionExecutor
         return true;
     }
 
-    public function displayAPI()
+    public function displayAPI(): array
     {
-        return $this->apiGet("/flux", []);
+        // Uniformisation des listes de sélection de type de dossier #2351
+        // getModuleListOrderByNom conservé par nom pour compatibilité appel API
+        return $this->objectInstancier->getInstance(ModuleListService::class)->getModuleListOrderByNom($this->id_u);
     }
 }