+++ test/PHPUnit/PastellTestCase.php
@@ -220,26 +220,6 @@ iparapheur_retour: Archive',
         return $this->getObjectInstancier()->getInstance(DonneesFormulaireFactory::class);
     }
 
-    /**
-     * @param $controllerName
-     * @param $id_u
-     * @return BaseAPIController
-     * @throws NotFoundException
-     */
-    protected function getAPIController($controllerName, $id_u)
-    {
-        /** @var  BaseAPIControllerFactory $factory */
-        $factory = $this->getObjectInstancier()->getInstance(BaseAPIControllerFactory::class);
-
-        if ($id_u) {
-            //FIXME : Faudrait pas que ca arrive...
-            /** @var Authentification $authentification */
-            $authentification = $this->objectInstancier->getInstance(Authentification::class);
-            $authentification->connexion('API', $id_u);
-        }
-        return $factory->getInstance($controllerName, $id_u);
-    }
-
     protected function getInternalAPI()
     {
         return $this->getInternalAPIAsUser(1);